+++ jest.config.js
@@ -1,6 +1,9 @@
 module.exports = {
     preset: 'ts-jest',
     testEnvironment: 'node',
+    transform: {
+        '^.+\\.ts$': ['ts-jest', { tsconfig: 'tsconfig.spec.json' }],
+    },
     collectCoverage: true,
     coverageDirectory: 'coverage',
     collectCoverageFrom: ['index.ts'],